Made some changes to the file structure and concerning the styling Please enter the commit message for your changes. Lines starting
