+++ src/components/atoms/index.js
@@ -1,2 +1,3 @@
 export * from './icons'
 export * from './buttons'
+export * from './input'
